pyproject.toml
[tool.poetry]
name = "photo-geocoder"
version = "0.1.0"
description = "A script to geocode images using EXIF data and rename them based on the city."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26"
# Keep a single Pillow entry – the latest version works with the rest of the stack
pillow = ">=10.0"
geopy = "^2.4.1"

# Torch / torchvision – specify the CPU build (most portable) or let the user choose via extras
torch = "^2.0"
torchvision = "^0.15"

scikit-image = ">=0.22"

[tool.poetry.dev-dependencies]
pytest = "^6.2"  # Optional, if you want to include testing libraries

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
